# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kronicle",
    "summary": "FastAPI-powered TimescaleDB microservice for storing time-series measurements",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/channels": {
      "get": {
        "tags": [
          "Read data"
        ],
        "summary": "list all available channels",
        "description": "Fetches metadata for all registered channels.\nEach entry includes schema, metadata, tags, and the number of available rows.\nNo data rows are returned in this endpoint.\nOptionally, filter by a name or tag_key/tag_value pair.",
        "operationId": "fetch_all_channels_metadata_api_v1_channels_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional name to filter by",
              "title": "Name"
            },
            "description": "Optional name to filter by"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Optional tags as comma-separated key:value pairs, e.g., color:red",
              "title": "Tags"
            },
            "description": "Optional tags as comma-separated key:value pairs, e.g., color:red"
          },
          {
            "name": "metadata",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Optional tags as comma-separated key:value pairs, e.g., color:red",
              "title": "Metadata"
            },
            "description": "Optional tags as comma-separated key:value pairs, e.g., color:red"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ResponsePayload"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ResponsePayload"
                    }
                  ],
                  "title": "Response Fetch All Channels Metadata Api V1 Channels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/channels/{channel_id}": {
      "get": {
        "tags": [
          "Read data"
        ],
        "summary": "Fetch metadata for a specific channel",
        "description": "Retrieves metadata for the specified `channel_id`, including schema, tags, and metadata.\nThe response also includes the number of rows stored for that channel\nbut does not include the row data itself.",
        "operationId": "fetch_channel_api_v1_channels__channel_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Channel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/channels/{channel_id}/rows": {
      "get": {
        "tags": [
          "Read data"
        ],
        "summary": "Fetch stored rows for a specific channel",
        "description": "Retrieves all stored time-series rows for the specified `channel_id`.\nFiltering, ordering, pagination, and column selection are supported via query parameters.\nThe response includes both metadata and data rows according to the channel's schema.",
        "operationId": "fetch_channel_rows_api_v1_channels__channel_id__rows_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Channel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsePayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelSchema": {
        "properties": {
          "column_types": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SchemaType"
            },
            "type": "object",
            "title": "Column Types",
            "description": "Column name -> SchemaType"
          },
          "db_to_usr": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Db To Usr",
            "description": "DB column name mapping -> User-defined column name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "column_types"
        ],
        "title": "ChannelSchema",
        "description": "ChannelSchema is a pure structural definition of a channel's timeseries layout.\n\nResponsibilities:\n- Define user-facing column names and their types.\n- Provide validation rules for rows.\n- Provide deterministic user\u2194internal column normalization.\n- Offer structural comparison (e.g. for migration/versioning).\n\nIt is:\n- Storage-agnostic (no table name, no SQL, no DB types).\n- Immutable (once created, structure does not change).\n- Safe to serialize and embed in API payloads (part of ChannelMetadata).\n\nIt does NOT:\n- Know about channel_id.\n- Generate SQL.\n- Know about row id or received_at.\n- Compare itself to a live database table.\n\nDatabase materialization is handled by ChannelTimeseries."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponsePayload": {
        "properties": {
          "channel_id": {
            "type": "string",
            "format": "uuid",
            "title": "Channel Id"
          },
          "channel_schema": {
            "$ref": "#/components/schemas/ChannelSchema"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "tags": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "boolean"
                    },
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    },
                    {
                      "type": "number"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "rows": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "columns": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {},
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "op_details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Op Details"
          }
        },
        "type": "object",
        "required": [
          "channel_id",
          "channel_schema"
        ],
        "title": "ResponsePayload",
        "description": "Payload returned in responses to the user.\n- issued_at: when the response was generated\n- available_rows: number of rows stored for this channel (None if unknown)"
      },
      "SchemaType": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "optional": {
            "type": "boolean",
            "title": "Optional",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SchemaType",
        "description": "Immutable, hashable type descriptor for channel columns.\n\nResponsibilities:\n- Stores canonical type name and optional flag.\n- Delegates validation and type mapping to SchemaRegistry.\n- Supports optional types via 'optional[...]' notation."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}